Refactor TrafficCamerasApp and Map components for improved structure
This commit moves the TrafficCamerasApp component to a new directory structure under Application and introduces a new CameraLocationMarker component for rendering camera markers on the map. The Map component is updated to utilize CameraLocationMarker, enhancing the organization and readability of the code. Additionally, the Home page is updated to reflect the new import path for TrafficCamerasApp.
