+++ nextjs/src/app/_components/Application/TrafficCamerasApp.tsx
